update docs, improve signals, other bugfixes (#430)
Changes:
- allow signals to interrupt softly deletions and other operations
- add `injected_filters` signal parameter for pre_delete
- update docs
- add `SkipOperation` exception
- add operation_skipped signal parameter
- improve tests for signals and caches
- cache fixes
- fix spurious failures in tests

---------

+++ tests/foreign_keys/test_ref_foreignkey.py
@@ -9,8 +9,8 @@ from tests.settings import DATABASE_URL
 
 pytestmark = pytest.mark.anyio
 
-database = DatabaseTestClient(DATABASE_URL, full_isolation=False)
-models = edgy.Registry(database=database)
+database = DatabaseTestClient(DATABASE_URL)
+models = edgy.Registry(database=edgy.Database(database, force_rollback=True))
 
 
 pydantic_version = ".".join(__version__.split(".")[:2])
@@ -64,16 +64,19 @@ class User(edgy.StrictModel):
 
 @pytest.fixture(autouse=True, scope="module")
 async def create_test_database():
-    await models.create_all()
-    yield
-    await models.drop_all()
+    # this creates and drops the database
+    async with database:
+        await models.create_all()
+        yield
+        if not database.drop:
+            await models.drop_all()
 
 
-@pytest.fixture(autouse=True)
-async def rollback_connections():
-    with database.force_rollback():
-        async with database:
-            yield
+@pytest.fixture(autouse=True, scope="function")
+async def rollback_transactions():
+    # this rolls back
+    async with models:
+        yield
 
 
 async def test_conversion():
